+++ src/core/base_classes/binomial_exponentiation_toolbox.h
@@ -120,14 +120,14 @@ namespace piranha
 				retval.add(tmp, args_tuple);
 				if (Op == power_op) {
 					for (size_t i = 1; i < n; ++i) {
-						tmp.mult_by(y - i + 1, args_tuple);
+						tmp.mult_by(y - (double)i + 1, args_tuple);
 						tmp.divide_by(i, args_tuple);
 						tmp.mult_by(XoverA, args_tuple);
 						retval.add(tmp, args_tuple);
 					}
 				} else {
 					for (size_t i = 1; i < n; ++i) {
-						tmp.mult_by(1 - i * y + y, args_tuple);
+						tmp.mult_by(1 - (double)i * y + y, args_tuple);
 						tmp.divide_by(y * i, args_tuple);
 						tmp.mult_by(XoverA, args_tuple);
 						retval.add(tmp, args_tuple);
